requirements.txt
setuptools>=80.0.0
grpcio>=1.67.1
grpcio-tools>=1.67.1
